+++ spec/trello/commands_spec.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+require 'trello/commands'
+
+RSpec.describe Kraken::Trello::Commands do
+  context '#cards' do
+    let(:client) { double }
+
+    let(:output) { capture(:stdout) { subject.cards(client) } }
+
+    it 'lists the cards for the trello board' do
+      expect(client).to receive(:list_cards).and_return([
+                                                          Kraken::Trello::Card.new(name: 'Build something new')
+                                                        ])
+
+      expected =  "Cards\n"
+      expected += "- Build something new\n"
+
+      expect(output).to eq(expected)
+    end
+  end
+end